Week nr, year, project og burger navn bliver under comments efter en create. Week nr & year kommer direkte ind. – ”Week” is included in week # column Deleting entries should now work properly. Searching in comments works. ReadOnly has been added to multiple “Create new” sub pages. Bundle(dynamiccrud) created for dynamicCrudPages.js (helps avoid refreshes) Clear button on time registration Time changed to float/double – should be update everywhere 13 new projects imported, import tool updated. Enforced standard on /DSA -> “_” turns to “-“. Feedback on create added on most pages. Union page feedback and clear on completion.

+++ TimeReg/TimeRegistration.cs
@@ -19,7 +19,7 @@ namespace TimeReg
         public int FK_OrderId { get; set; }
         public int FK_ProjectId { get; set; }
         public int FK_TaskId { get; set; }
-        public int Time { get; set; }
+        public double Time { get; set; }
         public System.DateTime Date { get; set; }
         public System.DateTime DateEntry { get; set; }
         public string Comment { get; set; }
