+++ TimeReg/VI_UserTimePerProject.cs
@@ -18,7 +18,7 @@ namespace TimeReg
         public string UserName { get; set; }
         public string NK_ZId { get; set; }
         public string Name { get; set; }
-        public Nullable<int> timeSum { get; set; }
+        public Nullable<double> timeSum { get; set; }
         public int FK_ProjectId { get; set; }
     }
 }
